aider: Added media queries to make the application responsive for iPhone screens.
# Aider chat conversation:

USER: Can you make this responsive for an iPhone screen?

ASSISTANT: To make the application responsive for an iPhone screen, we need to modify the CSS file to include media queries that will adjust the layout and design based on the screen size.

The files we need to modify are:
1. design5.css

Here are the changes we need to make:

1. Add media queries to adjust the layout for smaller screens.

Here are the specific changes:

```css
design5.css
<<<<<<< HEAD
=======
@media only screen and (max-width: 600px) {
    .container {
        width: 100%;
    }

    .accordion {
        width: 100%;
    }

    .tally-display {
        width: 100%;
        font-size: 2em;
    }
}
>>>>>>> updated
```

This media query will apply the enclosed styles only when the screen width is 600px or less, which covers iPhone screens. The changes include setting the width of the container and accordion to 100% to utilize the full width of the screen, and reducing the font size of the tally display to better fit smaller screens.