aider: Set the width of the container to 100%.
# Aider chat conversation:

USER: Can you make it span the screen?

ASSISTANT: To make the application span the entire screen, we need to modify the CSS file to set the width of the container to 100%.

The files we need to modify are:
1. design5.css

Here are the changes we need to make:

1. Set the width of the container to 100%.

Here are the specific changes:

```css
design5.css
<<<<<<< HEAD
.container {
    display: flex;
    flex-direction: column;
    align-items: center;
    width: 75%;
}
=======
.container {
    display: flex;
    flex-direction: column;
    align-items: center;
    width: 100%;
}
>>>>>>> updated
```

This change will make the container span the entire width of the screen.

USER: InvalidEditBlock: edit failed!

design5.css does not contain the *exact sequence* of HEAD lines you specified.
Try again.
DO NOT skip blank lines, comments, docstrings, etc!
The HEAD block needs to be EXACTLY the same as the lines in design5.css with nothing missing!

design5.css does not contain these 6 exact lines in a row:
```
.container {
    display: flex;
    flex-direction: column;
    align-items: center;
    width: 75%;
}
```

ASSISTANT: I apologize for the confusion. It seems there was a mistake in identifying the exact lines to be replaced. Let's try again.

The correct lines to be replaced in the design5.css file are:

```css
design5.css
<<<<<<< HEAD
.container {
    display: flex;
    flex-direction: column;
    align-items: center;
}
=======
.container {
    display: flex;
    flex-direction: column;
    align-items: center;
    width: 100%;
}
>>>>>>> updated
```

This change will make the container span the entire width of the screen.