aider: Set the width of the `.accordion-btn` to 100% and increase the font size within the media query.
# Aider chat conversation:

USER: On my phone the menu spans the screen but the menu text does not. It is very small and scrunched up to the left.

ASSISTANT: To make the menu text span the entire width of the screen and increase its size for better readability on smaller screens, we need to modify the CSS file.

The files we need to modify are:
1. design5.css

Here are the changes we need to make:

1. Set the width of the `.accordion-btn` to 100% and increase the font size within the media query.

Here are the specific changes:

```css
design5.css
<<<<<<< HEAD
@media only screen and (max-width: 600px) {
    .container {
        width: 100%;
    }

    .accordion {
        width: 100%;
    }

    .tally-display {
        width: 100%;
        font-size: 2em;
    }
}
=======
@media only screen and (max-width: 600px) {
    .container {
        width: 100%;
    }

    .accordion {
        width: 100%;
    }

    .accordion-btn {
        width: 100%;
        font-size: 1.5em;
    }

    .tally-display {
        width: 100%;
        font-size: 2em;
    }
}
>>>>>>> updated
```

This change will make the menu text span the entire width of the screen and increase its size when the screen width is 600px or less, which covers iPhone screens.